# apps/api/internal/api/api.yaml
openapi: 3.0.3
info:
  title: BlogPost Platform API
  version: 1.0.0
  description: >
    This API allows users to manage blog posts, comments, and user accounts on a blogging platform.

servers:
- url: https://example.com/api/v1
  description: Production server
- url: http://localhost:8080/api/v1
  description: Local server

tags:
- name: Posts
  description: Endpoints related to blog posts
- name: Comments
  description: Endpoints related to post comments
- name: Users
  description: Endpoints related to user management

paths:
  /posts:
    get:
      tags:
      - Posts
      summary: List all posts
      operationId: listPosts
      parameters:
      - in: query
        name: page
        schema:
          type: integer
        description: Page number for pagination
      - in: query
        name: pageSize
        schema:
          type: integer
          default: 10
        description: Number of posts per page
      responses:
        '200':
          description: A list of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'

    post:
      tags:
      - Posts
      summary: Create a new post
      operationId: createPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostInput'
      responses:
        '201':
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Invalid input

  /posts/{postId}:
    get:
      tags:
      - Posts
      summary: Get a single post
      operationId: getPost
      parameters:
      - name: postId
        in: path
        required: true
        schema:
          type: string
        description: Unique ID of the post to retrieve
      responses:
        '200':
          description: The requested post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          description: Post not found

    put:
      tags:
      - Posts
      summary: Update an existing post
      operationId: updatePost
      parameters:
      - name: postId
        in: path
        required: true
        schema:
          type: string
        description: ID of the post to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostInput'
      responses:
        '200':
          description: Post updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Invalid input
        '404':
          description: Post not found

    delete:
      tags:
      - Posts
      summary: Delete a post
      operationId: deletePost
      parameters:
      - name: postId
        in: path
        required: true
        schema:
          type: string
        description: ID of the post to delete
      responses:
        '204':
          description: Post deleted successfully
        '404':
          description: Post not found

  /posts/{postId}/comments:
    get:
      tags:
      - Comments
      summary: List comments for a post
      operationId: listComments
      parameters:
      - name: postId
        in: path
        required: true
        schema:
          type: string
        description: Post ID for which to retrieve comments
      responses:
        '200':
          description: A list of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'

    post:
      tags:
      - Comments
      summary: Create a comment on a specific post
      operationId: createComment
      parameters:
      - name: postId
        in: path
        required: true
        schema:
          type: string
        description: Post ID to which this comment belongs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentInput'
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: Invalid input

  /posts/{postId}/comments/{commentId}:
    get:
      tags:
      - Comments
      summary: Retrieve a specific comment
      operationId: getComment
      parameters:
      - name: postId
        in: path
        required: true
        schema:
          type: string
        description: Post ID to which the comment belongs
      - name: commentId
        in: path
        required: true
        schema:
          type: string
        description: Comment ID
      responses:
        '200':
          description: The requested comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '404':
          description: Comment not found

    put:
      tags:
      - Comments
      summary: Update a specific comment
      operationId: updateComment
      parameters:
      - name: postId
        in: path
        required: true
        schema:
          type: string
        description: Post ID
      - name: commentId
        in: path
        required: true
        schema:
          type: string
        description: Comment ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentInput'
      responses:
        '200':
          description: Comment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '404':
          description: Comment not found

    delete:
      tags:
      - Comments
      summary: Delete a specific comment
      operationId: deleteComment
      parameters:
      - name: postId
        in: path
        required: true
        schema:
          type: string
        description: Post ID
      - name: commentId
        in: path
        required: true
        schema:
          type: string
        description: Comment ID
      responses:
        '204':
          description: Comment deleted successfully
        '404':
          description: Comment not found

  /users:
    get:
      tags:
      - Users
      summary: Get all users
      operationId: listUsers
      parameters:
      - in: query
        name: page
        schema:
          type: integer
        description: Page number for pagination
      - in: query
        name: pageSize
        schema:
          type: integer
          default: 10
        description: Number of users per page
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'

    post:
      tags:
      - Users
      summary: Create a new user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid input

  /users/{userId}:
    get:
      tags:
      - Users
      summary: Get user by ID
      operationId: getUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
        description: ID of the user to retrieve
      responses:
        '200':
          description: The requested user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found

    patch:
      tags:
      - Users
      summary: Update user by ID
      operationId: updateUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
        description: ID of the user to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found

    delete:
      tags:
      - Users
      summary: Delete user by ID
      operationId: deleteUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
        description: ID of the user to delete
      responses:
        '204':
          description: User deleted successfully
        '404':
          description: User not found

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Post:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        authorId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    PostInput:
      type: object
      required:
      - title
      - content
      properties:
        title:
          type: string
        content:
          type: string

    Comment:
      type: object
      properties:
        id:
          type: string
        postId:
          type: string
        authorId:
          type: string
        text:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CommentInput:
      type: object
      required:
      - text
      properties:
        text:
          type: string

    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    UserInput:
      type: object
      required:
      - username
      - email
      properties:
        username:
          type: string
        email:
          type: string
          format: email

security:
- bearerAuth: []